pyproject.toml
[tool.poetry]
name = "hand_tracking"
version = "0.1.0"
description = ""
authors = ["Marcus Braga <mvbraga@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
setuptools = "^65.6.3"
opencv-python = "^4.6.0"
openai = "^0.25.0"
beautifulsoup4 = "^4.11.1"
Pillow = "^9.3.0"
pygame = "^2.1.2"
whisper = "^1.1.10"
invoke = "^1.7.3"
jupyter = "^1.0.0"
python-decouple = "^3.6"
mediapipe = "^0.9.0"
pycaw = "^20220416"
transformers = "^4.25.1"
pre-commit = "^2.20.0"
flake8 = "^6.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
skip_glob = "data/*"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
line_length = 88

[tool.black]
py37 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
)/
'''

[tool.flake8]
max-line-length = 120
exclude = "data/*"
max-complexity = 10
ignore = "E231"

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true
